rtpsession: Calculate RTCP bandwidth as a fraction of the RTP bandwidth
Calculate the RTCP bandwidth to be a fraction of the RTP bandwidth if it is
specified as a value between 0 and 1.

+++ gst/rtpmanager/rtpsession.h
@@ -166,7 +166,7 @@ struct _RTPSession {
   /* bandwidths */
   gboolean     recalc_bandwidth;
   guint        bandwidth;
-  guint        rtcp_bandwidth;
+  gdouble      rtcp_bandwidth;
   guint        rtcp_rr_bandwidth;
   guint        rtcp_rs_bandwidth;
